ART-56 Moved Networking tools to its own folder
