+++ app/src/test/java/com/sefford/artdrian/test/networking/MockEngine.kt
@@ -1,4 +1,4 @@
-package com.sefford.artdrian.test
+package com.sefford.artdrian.test.networking
 
 import io.ktor.client.engine.mock.MockRequestHandleScope
 import io.ktor.client.engine.mock.respond
